Introduce a draft implementation of CgSpringTestClassConstructor
