+++ utbot-framework/src/main/kotlin/org/utbot/framework/codegen/tree/CgSimpleTestClassConstructor.kt
@@ -1,13 +1,11 @@
 package org.utbot.framework.codegen.tree
 
-import org.utbot.framework.UtSettings
 import org.utbot.framework.codegen.domain.Junit4
 import org.utbot.framework.codegen.domain.Junit5
 import org.utbot.framework.codegen.domain.ParametrizedTestSource
 import org.utbot.framework.codegen.domain.TestNg
 import org.utbot.framework.codegen.domain.builtin.TestClassUtilMethodProvider
 import org.utbot.framework.codegen.domain.context.CgContext
-import org.utbot.framework.codegen.domain.models.CgAuxiliaryClass
 import org.utbot.framework.codegen.domain.models.CgClassBody
 import org.utbot.framework.codegen.domain.models.CgMethod
 import org.utbot.framework.codegen.domain.models.CgMethodTestSet
@@ -17,36 +15,16 @@ import org.utbot.framework.codegen.domain.models.CgRegion
 import org.utbot.framework.codegen.domain.models.CgSimpleRegion
 import org.utbot.framework.codegen.domain.models.CgStaticsRegion
 import org.utbot.framework.codegen.domain.models.CgTestMethod
-import org.utbot.framework.codegen.domain.models.CgTestMethodCluster
-import org.utbot.framework.codegen.domain.models.CgTripleSlashMultilineComment
-import org.utbot.framework.codegen.domain.models.CgUtilEntity
-import org.utbot.framework.codegen.domain.models.CgUtilMethod
 import org.utbot.framework.codegen.domain.models.SimpleTestClassModel
-import org.utbot.framework.codegen.renderer.importUtilMethodDependencies
-import org.utbot.framework.codegen.reports.TestsGenerationReport
-import org.utbot.framework.codegen.tree.CgComponents.clearContextRelatedStorage
-import org.utbot.framework.codegen.tree.CgComponents.getMethodConstructorBy
-import org.utbot.framework.plugin.api.ClassId
-import org.utbot.framework.plugin.api.MethodId
 import org.utbot.framework.plugin.api.UtExecutionSuccess
-import org.utbot.framework.plugin.api.UtMethodTestSet
 import org.utbot.framework.plugin.api.UtSymbolicExecution
-import org.utbot.framework.plugin.api.util.description
 import org.utbot.framework.plugin.api.util.humanReadableName
 import org.utbot.fuzzer.UtFuzzedExecution
 
 /**
  * This test class constructor is used for pure Java/Kotlin applications.
  */
-open class CgSimpleTestClassConstructor(context: CgContext): CgTestClassConstructorBase<SimpleTestClassModel>(context) {
-
-    init {
-        clearContextRelatedStorage()
-    }
-
-    override val methodConstructor = getMethodConstructorBy(context)
-
-    val testsGenerationReport = TestsGenerationReport()
+open class CgSimpleTestClassConstructor(context: CgContext): CgAbstractTestClassConstructor<SimpleTestClassModel>(context) {
 
     override fun constructTestClassBody(testClassModel: SimpleTestClassModel): CgClassBody =
         buildClassBody(currentTestClass) {
@@ -100,7 +78,7 @@ open class CgSimpleTestClassConstructor(context: CgContext): CgTestClassConstruc
             }
         }
 
-    private fun constructTestSet(testSet: CgMethodTestSet): List<CgRegion<CgMethod>>? {
+    override fun constructTestSet(testSet: CgMethodTestSet): List<CgRegion<CgMethod>>? {
         val regions = mutableListOf<CgRegion<CgMethod>>()
 
         if (testSet.executions.any()) {
@@ -137,47 +115,6 @@ open class CgSimpleTestClassConstructor(context: CgContext): CgTestClassConstruc
         }
 
         return testSets
-    }
-
-    private fun processFailure(testSet: CgMethodTestSet, failure: Throwable) {
-        codeGenerationErrors
-            .getOrPut(testSet) { mutableMapOf() }
-            .merge(failure.description, 1, Int::plus)
-    }
-
-    private fun createTest(
-        testSet: CgMethodTestSet,
-        regions: MutableList<CgRegion<CgMethod>>
-    ) {
-        val (methodUnderTest, _, clustersInfo) = testSet
-
-        for ((clusterSummary, executionIndices) in clustersInfo) {
-            val currentTestCaseTestMethods = mutableListOf<CgTestMethod>()
-            emptyLineIfNeeded()
-            val (checkedRange, needLimitExceedingComments) = if (executionIndices.last  - executionIndices.first >= UtSettings.maxTestsPerMethodInRegion) {
-                IntRange(executionIndices.first, executionIndices.first + (UtSettings.maxTestsPerMethodInRegion - 1).coerceAtLeast(0)) to true
-            } else {
-                executionIndices to false
-            }
-
-            for (i in checkedRange) {
-                currentTestCaseTestMethods += methodConstructor.createTestMethod(methodUnderTest, testSet.executions[i])
-            }
-
-            val comments = listOf("Actual number of generated tests (${executionIndices.last - executionIndices.first}) exceeds per-method limit (${UtSettings.maxTestsPerMethodInRegion})",
-                "The limit can be configured in '{HOME_DIR}/.utbot/settings.properties' with 'maxTestsPerMethod' property")
-
-            val clusterHeader = clusterSummary?.header
-            var clusterContent = clusterSummary?.content
-                ?.split('\n')
-                ?.let { CgTripleSlashMultilineComment(if (needLimitExceedingComments) {it.toMutableList() + comments} else {it}) }
-            if (clusterContent == null && needLimitExceedingComments) {
-                clusterContent = CgTripleSlashMultilineComment(comments)
-            }
-            regions += CgTestMethodCluster(clusterHeader, clusterContent, currentTestCaseTestMethods)
-
-            testsGenerationReport.addTestsByType(testSet, currentTestCaseTestMethods)
-        }
     }
 
     private fun createParametrizedTestAndDataProvider(
@@ -245,79 +182,5 @@ open class CgSimpleTestClassConstructor(context: CgContext): CgTestClassConstruc
 
         return testMethods
     }
-
-    /**
-     * This method collects a list of util entities (methods and classes) needed by the class.
-     * By the end of the test method generation [requiredUtilMethods] may not contain all the needed.
-     * That's because some util methods may not be directly used in tests, but they may be used from other util methods.
-     * We define such method dependencies in [MethodId.methodDependencies].
-     *
-     * Once all dependencies are collected, required methods are added back to [requiredUtilMethods],
-     * because during the work of this method they are being removed from this list, so we have to put them back in.
-     *
-     * Also, some util methods may use some classes that also need to be generated.
-     * That is why we collect information about required classes using [MethodId.classDependencies].
-     *
-     * @return a list of [CgUtilEntity] representing required util methods and classes (including their own dependencies).
-     */
-    private fun collectUtilEntities(): List<CgUtilEntity> {
-        val utilMethods = mutableListOf<CgUtilMethod>()
-        // Some util methods depend on other util methods or some auxiliary classes.
-        // Using this loop we make sure that all the util method dependencies are taken into account.
-        val requiredClasses = mutableSetOf<ClassId>()
-        while (requiredUtilMethods.isNotEmpty()) {
-            val method = requiredUtilMethods.first()
-            requiredUtilMethods.remove(method)
-            if (method.name !in existingMethodNames) {
-                utilMethods += CgUtilMethod(method)
-                // we only need imports from util methods if these util methods are declared in the test class
-                if (utilMethodProvider is TestClassUtilMethodProvider) {
-                    importUtilMethodDependencies(method)
-                }
-                existingMethodNames += method.name
-                requiredUtilMethods += method.methodDependencies()
-                requiredClasses += method.classDependencies()
-            }
-        }
-        // Collect all util methods back into requiredUtilMethods.
-        // Now there will also be util methods that weren't present in requiredUtilMethods at first,
-        // but were needed for the present util methods to work.
-        requiredUtilMethods += utilMethods.map { method -> method.id }
-
-        val auxiliaryClasses = requiredClasses.map { CgAuxiliaryClass(it) }
-
-        return utilMethods + auxiliaryClasses
-    }
-
-    /**
-     * If @receiver is an util method, then returns a list of util method ids that @receiver depends on
-     * Otherwise, an empty list is returned
-     */
-    private fun MethodId.methodDependencies(): List<MethodId> = when (this) {
-        createInstance -> listOf(getUnsafeInstance)
-        deepEquals -> listOf(arraysDeepEquals, iterablesDeepEquals, streamsDeepEquals, mapsDeepEquals, hasCustomEquals)
-        arraysDeepEquals, iterablesDeepEquals, streamsDeepEquals, mapsDeepEquals -> listOf(deepEquals)
-        buildLambda, buildStaticLambda -> listOf(
-            getLookupIn, getSingleAbstractMethod, getLambdaMethod,
-            getLambdaCapturedArgumentTypes, getInstantiatedMethodType, getLambdaCapturedArgumentValues
-        )
-        else -> emptyList()
-    }
-
-    /**
-     * If @receiver is an util method, then returns a list of auxiliary class ids that @receiver depends on.
-     * Otherwise, an empty list is returned.
-     */
-    private fun MethodId.classDependencies(): List<ClassId> = when (this) {
-        buildLambda, buildStaticLambda -> listOf(capturedArgumentClass)
-        else -> emptyList()
-    }
-
-    /**
-     * Engine errors + codegen errors for a given [UtMethodTestSet]
-     */
-    protected val CgMethodTestSet.allErrors: Map<String, Int>
-        get() = errors + codeGenerationErrors.getOrDefault(this, mapOf())
-
 }
