+++ utbot-framework/src/main/kotlin/org/utbot/framework/codegen/tree/CgSpringTestClassConstructor.kt
@@ -1,15 +1,64 @@
 package org.utbot.framework.codegen.tree
 
+import org.utbot.framework.codegen.domain.builtin.TestClassUtilMethodProvider
 import org.utbot.framework.codegen.domain.context.CgContext
 import org.utbot.framework.codegen.domain.models.CgClassBody
+import org.utbot.framework.codegen.domain.models.CgMethod
+import org.utbot.framework.codegen.domain.models.CgMethodTestSet
+import org.utbot.framework.codegen.domain.models.CgMethodsCluster
+import org.utbot.framework.codegen.domain.models.CgRegion
+import org.utbot.framework.codegen.domain.models.CgStaticsRegion
 import org.utbot.framework.codegen.domain.models.SpringTestClassModel
 
-class CgSpringTestClassConstructor(context: CgContext): CgTestClassConstructorBase<SpringTestClassModel>(context) {
-
-    override val methodConstructor: CgMethodConstructor
-        get() = TODO("Not yet implemented")
+class CgSpringTestClassConstructor(context: CgContext): CgAbstractTestClassConstructor<SpringTestClassModel>(context) {
 
     override fun constructTestClassBody(testClassModel: SpringTestClassModel): CgClassBody {
-        TODO("Not yet implemented")
+        return buildClassBody(currentTestClass) {
+
+            // TODO: support inner classes here
+
+            // TODO: create class variables with Mock/InjectMock annotations using testClassModel
+            fields += mutableListOf()
+
+            // TODO: create beforeEach/ afterEach methods
+            // Requires new implementation of CgMethod for test framework builtins
+
+            for (testSet in testClassModel.methodTestSets) {
+                updateCurrentExecutable(testSet.executableId)
+                val currentMethodUnderTestRegions = constructTestSet(testSet) ?: continue
+                val executableUnderTestCluster = CgMethodsCluster(
+                    "Test suites for executable $currentExecutable",
+                    currentMethodUnderTestRegions
+                )
+                methodRegions += executableUnderTestCluster
+            }
+
+            if (currentTestClass == outerMostTestClass) {
+                val utilEntities = collectUtilEntities()
+                // If utilMethodProvider is TestClassUtilMethodProvider, then util entities should be declared
+                // in the test class. Otherwise, util entities will be located elsewhere (e.g. another class).
+                if (utilMethodProvider is TestClassUtilMethodProvider && utilEntities.isNotEmpty()) {
+                    staticDeclarationRegions += CgStaticsRegion("Util methods", utilEntities)
+                }
+            }
+        }
+    }
+
+    override fun constructTestSet(testSet: CgMethodTestSet): List<CgRegion<CgMethod>>? {
+        val regions = mutableListOf<CgRegion<CgMethod>>()
+
+        if (testSet.executions.any()) {
+            runCatching {
+                createTest(testSet, regions)
+            }.onFailure { e -> processFailure(testSet, e) }
+        }
+
+        val errors = testSet.allErrors
+        if (errors.isNotEmpty()) {
+            regions += methodConstructor.errorMethod(testSet.executableId, errors)
+            testsGenerationReport.addMethodErrors(testSet, errors)
+        }
+
+        return if (regions.any()) regions else null
     }
 }
